+++ tests/testthat/test-plot_helpers.R
@@ -1,6 +1,6 @@
 # plot_pop_helper ----
 test_that("plot_pop_helper snapshot", {
-  library(tibble)
+  suppressWarnings(library(tibble))
   set.seed(1)
   x <- rweibull(n = 100, shape = 1, scale = 20000)
   param_tbl <- tibble(param_1 = 20000, param_2 = 1)